docs: HttpClient revamp
