--- app/ng-framework-modules-category/fps-meter/fps-meter-usage/article.md
@@ -1,8 +1,0 @@
-Import FPS Meter Module
-<snippet id='fps-meter-module-import'/>
-
-Add Callback method and start logging
-<snippet id='start-fps-meter'/>
-
-Remove Callback method and stop logging
-<snippet id='stop-fps-meter'/>
