--- app/ng-framework-modules-category/fps-meter/fps-meter-usage/fps-meter.html
@@ -1,9 +1,0 @@
-<StackLayout sdkExampleTitle sdkToggleNavButton>
-    <GridLayout rows="auto auto" columns="* *">
-        <Label class="p-15" row="0" col="0" text="FPS" textWrap="true"></Label>
-        <Label class="p-15" row="0" col="1" [text]="fps"></Label>
-        <Label class="p-15" row="1" col="0" text="minFPS" textWrap="true"></Label>
-        <Label class="p-15" row="1" col="1" [text]="minfps"></Label>
-    </GridLayout>
-    <Button class="btn btn-primary btn-active" [text]="fpslabeltitle" (tap)="toogleMeter()"></Button>
-</StackLayout>
