+++ app/ng-framework-modules-category/fps-meter/overview.md
@@ -1,1 +1,1 @@
-Logging frames-per-second statistics for your app requires the "fps-meter" module.
+Logging frames-per-second statistics for your app requires the `fps-meter` module.
