+++ app/ng-framework-modules-category/fps-meter/usage/article.md
@@ -1,0 +1,2 @@
+Import FPS Meter Module
+<snippet id='fps-meter-module-usage'/>
