+++ app/ng-framework-modules-category/fps-meter/usage/usage.component.html
@@ -1,0 +1,5 @@
+<GridLayout rows="*, *" columns="*" sdkExampleTitle sdkToggleNavButton>
+    <Button row="0" text="Start FPSMeter" (tap)="startFPSMeter()"></Button>
+    <Button row="1" text="Stop FPSMeter" (tap)="stopFPSMeter()"></Button>
+</GridLayout>
+
