+++ app/ng-framework-modules-category/http/end.md
@@ -1,2 +1,9 @@
+## Methods
 
-**API Reference for the** [HTTP Class](http://docs.nativescript.org/api-reference/modules/_http_.html)
+Refer to [the Angular's official documentation](https://angular.io/guide/http) for more details on `HttpClient` module.
+
+## API References
+
+| Name     | Type    | 
+|----------|---------|
+| [HttpClient](https://angular.io/guide/http) | `Module` |
