--- app/ng-framework-modules-category/http/http-delete/article.md
@@ -1,5 +1,0 @@
-Using the DELETE method by creating a service file to keep the HTTP logic separated from the component file.
-<snippet id='http-delete-service'/>
-
-Provide the service in the component (or in the related `NgModule` if the service should be reused).
-<snippet id='http-delete-component'/>
