--- app/ng-framework-modules-category/http/http-put/article.md
@@ -1,5 +1,0 @@
-Using the PUT method by creating a service file to keep the HTTP logic separated from the component file.
-<snippet id='http-put-service'/>
-
-Provide the service in the component (or in the related `NgModule` if the service should be reused).
-<snippet id='http-put-component'/>
